aita
AITA for telling my friend it’s her fault for getting married and having kids late because the world won’t wait on her now

I (39F) have a 6 person girl group since college (37-39F) and that includes Mary (38F). We’ve been close throughout the years and have been at milestone events for each other. Mary just had a baby and is completely fitting the crazy new mother stereotype.

In college, Mary has always been someone who had to make it known that she was unique/different from the rest of us which wasn’t as draining then as it has become now. For starters, all other women in our circle, got married between the ages of 22-27 and we all have multiple kids. So the 5 of us were able to experience those milestones alongside one another and got closer as we shared similar lifestyles.

Mary was very adamant on not settling until her 30s because she wanted to travel and have different experiences which we all supported. Regardless, she would continue to make comments about how she’s so lucky unlike us because we’re “tied down with husbands and babies”. I think this is where she grew resentment towards us because we were in different places in life and she was upset we couldn’t have our group be similar to how it was in college.

Then into our mid 30s it became a whole saga of she’s getting older and can’t find a husband because all the “good men” are married or divorced with kids. When she finally got married, many could not attend because it was a destination event and child-free during Covid. This caused a fight because she said how she was there for us during our weddings but we couldn’t put aside a week for her. We had all told her how we wished we could, but it simply was not financially feasible and didn’t logistically work with our kids. But she just refused to hear us out and was simply so inconsiderate about our lives and families, saying we were horrible friends.

Now, Mary just gave birth to her first child and I was very excited for her. The only issue is that she moved from our state to a very remote place that’s only accessible by a 6hr car ride. Her baby is 6mo old and none of us have been able to go up to visit her. I think she’s been having a wrong idea of what a “village” is and has essentially demanded in our groupchat that we come up for the holidays and help her out because she’s having a hard time adjusting to mom life. But this would entail we all take a week off, arrange childcare, figure out transportation, and book hotels during the holidays. It’s gotten to the point where she’s posting cryptic messages on Facebook bashing “fake friends” who won’t be there for her. As much as I wish I could, I cannot physically support her in the way she needs me to do in this stage of life. It would have been completely different if she still lived in our city and this was earlier in life when we had less commitments/priorities. So I told her this and that if she was hoping for this big village and constant support, she should have thought about that when planning out her life because we can’t all just pause our lives for her. So AITA?